Update README.md
Suppose you are given a row of n doors fd1; d2; :::; dng, each of which can either be locked or unlocked. Your goal is to have exactly S secured doors. The issue is that not all locked doors are secure. A door is secure if and only if:

This door (di) is locked AND

This door is the first door (i.e., d1) OR The door to this one’s left (di 1) is also locked

Develop an algorithm that, given the number of doors n and a value S, counts the total number of unique ways the doors can be locked to achieve EXACTLY S secure doors.

You are going skiing this weekend and wish to ski once down every unique run the moun-tain has to offer. The mountain has n runs given to you in a list called R = fr1; r2; :::; rng where each ri is a positive integer denoting the number of minutes it will take to ski that run and get back up the lift to the top of the mountain. You want to ski each run in the minimum number of days possible. Each day has L minutes of skiing available. Obviously, you must fit each run into a single day (you can’t split a run over two days). You also wish to ski the runs in order.